Refactoring file system
