+++ src/Interfaces/MainInterfaces.ts
