+++ src/Routes/MainPage/mainPage.scss
@@ -6,7 +6,7 @@
 .main-content {
   width: 100%;
   height: 100%;
-  background-image: url(../../assets/bgmain.jpg);
+  background-image: url(../../Assets/JPG/main-bg.jpg);
   background-repeat: repeat;
   background-size: cover;
 }
