+++ src/Routes/PlayerInfo/PlayerInfo.tsx
@@ -1,18 +1,18 @@
 import React from 'react'
 
 import './playerInfo.scss'
-import { useAppSelector } from '../../PreTypedHooks'
-import PlayerAvatar from '../../components/Blocks/PlayerAvatar/PlayerAvatar'
-import SocialList from '../../components/Blocks/SocialList/SocialList'
-import StatsLine from '../../components/Blocks/StatsLine/StatsLine'
-import VerticalLine from '../../components/Custom/VerticalLine/VerticalLine'
-import Heading from '../../components/Blocks/Heading/Heading'
-import PlainText from '../../components/Blocks/PlainText/PlainText'
-import Devider from '../../components/Custom/Devider/Devider'
-import PlayerFollow from '../../components/Blocks/PlayerFollow/PlayerFollow'
-import PlayersList from '../../components/Blocks/PlayersList/PlayersList'
-import ParalaxHeading from '../../components/Blocks/ParalaxHeading/ParalaxHeading'
-import Footer from '../../components/Sections/Footer/Footer'
+import { useAppSelector } from '../../Hooks/PreTypedHooks'
+import PlayerAvatar from '../../Components/Blocks/PlayerAvatar/PlayerAvatar'
+import SocialList from '../../Components/Blocks/SocialList/SocialList'
+import StatsLine from '../../Components/Blocks/StatsLine/StatsLine'
+import VerticalLine from '../../Components/Custom/VerticalLine/VerticalLine'
+import Heading from '../../Components/Blocks/Heading/Heading'
+import PlainText from '../../Components/Blocks/PlainText/PlainText'
+import Devider from '../../Components/Custom/Devider/Devider'
+import PlayerFollow from '../../Components/Blocks/PlayerFollow/PlayerFollow'
+import PlayersList from '../../Components/Blocks/PlayersList/PlayersList'
+import ParalaxHeading from '../../Components/Blocks/ParalaxHeading/ParalaxHeading'
+import Footer from '../../Components/Sections/Footer/Footer'
 
 function PlayerInfo() {
   const currentPlayer = useAppSelector((state) => state.main.currentPlayer)
