+++ src/Routes/PlayerInfo/playerInfo.scss
@@ -13,7 +13,7 @@
     position: relative;
     width: 100%;
     height: auto;
-    background-image: url(../../assets/bluebg.jpg);
+    background-image: url(../../Assets/JPG/portal-bg.jpg);
     background-position: top center;
     background-repeat: no-repeat;
     background-size: cover;
@@ -27,7 +27,7 @@
       transform: translate(-50%, 0);
       width: 110%;
       height: 5rem;
-      background-image: url(../../assets/bottom-shape-round.svg);
+      background-image: url(../../Assets/SVG/bottom-shape-round.svg);
     }
 
     &--overlay {
@@ -99,7 +99,7 @@
 
   &__related-teammates {
     padding-top: var(--space-lg);
-    background-image: url(../../assets/aboutbg2.jpg);
+    background-image: url(../../Assets/JPG/portal-bg.jpg);
     @include cover-background;
     @include padding-top-lg;
     padding-bottom: 5rem;
