+++ src/components/Blocks/ParalaxHeading/paralaxHeading.scss
@@ -7,7 +7,7 @@
   width: 100%;
   height: 100%;
   @include gutter-xl;
-  background-image: url(../../../assets/second-bg.jpg);
+  background-image: url(../../../Assets/JPG/portal-bg.jpg);
   @include cover-background;
   background-attachment: fixed;
   position: relative;
