+++ src/components/Custom/CubePreloader/CubePreloader.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import './cubePreloader.scss'
-import preloader1 from '../../../assets/Preloaders/preloader1.gif'
+import preloader1 from '../../../Assets/Preloaders/preloader1.gif'
 
 function Preloader() {
   return (
