+++ src/components/Sections/Footer/footer.scss
@@ -6,7 +6,7 @@
   width: 100%;
   height: 100%;
   padding: 2rem;
-  background-image: url(../../../assets/aboutbg2.jpg);
+  background-image: url(../../../Assets/JPG/portal-bg.jpg);
   @include cover-background;
   position: relative;
   z-index: 0;
